implemented input actions

+++ Game/Components/LevelGridComponent.hpp
@@ -15,14 +15,14 @@ namespace dae
             std::unordered_map<char, std::string> textures);
 
         virtual void Render() const override;
-        bool IsOnPlatform(glm::ivec2 const &topleft, glm::ivec2 size);
-        bool IsOnLadder(glm::ivec2 const &topleft, glm::ivec2 size);
+        bool IsOnPlatform(glm::vec2 const &topleft, glm::vec2 size);
+        bool IsOnLadder(glm::vec2 const &topleft, glm::vec2 size);
 
       private:
         glm::ivec2 m_GridSize{};
         std::vector<std::vector<char>> m_Grid{};
         std::unordered_map<char, SDL_Texture *> m_TexturesLoaded;
 
-        char GetTile(glm::ivec2 const &pos);
+        char GetTile(glm::vec2 const &pos);
     };
 } // namespace dae
